requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-pptx>=0.6.23
transformers>=4.36.2
sentence-transformers>=2.3.1
torch>=2.9.0
accelerate>=0.26.1
bitsandbytes>=0.41.3
pydantic>=2.5.3
python-multipart>=0.0.6
numpy>=1.24.3
sentencepiece>=0.1.99
protobuf>=3.20.3
